Add plotting examples
