+++ examples/basics/scene/grid.py
@@ -1,1 +1,1 @@
-../../scene/grid.py
+../../scene/grid_layout/grid.py
