+++ examples/gloo/spatial_filters.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
+# vispy: gallery 2
 """
 Spatial Filtering
 =================
