+++ examples/scene/grid_layout/README.rst
@@ -1,0 +1,4 @@
+Grid Layouts
+============
+
+Examples of using the GridWidget to layout elements in a SceneCanvas.
