+++ examples/scene/grid_layout/grid.py
