+++ examples/scene/isosurface.py
@@ -3,7 +3,7 @@
 # Copyright (c) Vispy Development Team. All Rights Reserved.
 # Distributed under the (new) BSD License. See LICENSE.txt for more info.
 # -----------------------------------------------------------------------------
-
+# vispy: gallery 2
 """
 Isosurface Visual
 =================
