+++ examples/scene/polygon.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
 # Copyright (c) Vispy Development Team. All Rights Reserved.
 # Distributed under the (new) BSD License. See LICENSE.txt for more info.
-
+# vispy: gallery 2
 """
-Scene Visuals
+Shape Visuals
 =============
 
 Demonstration of PolygonVisual, EllipseVisual, RectangleVisual
