+++ examples/scene/shape_draw.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 # Copyright (c) 2018, Felix Schill.
 # Distributed under the (new) BSD License. See LICENSE.txt for more info.
+# vispy: gallery 2
 """
 Draw and Edit Shapes with Mouse
 ===============================
