+++ examples/scene/text.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
+# vispy: gallery 2
 # -----------------------------------------------------------------------------
 # Copyright (c) Vispy Development Team. All Rights Reserved.
 # Distributed under the (new) BSD License. See LICENSE.txt for more info.
 # -----------------------------------------------------------------------------
-
 """
 Text in a Scene and ViewBox
 ===========================
